requirements.txt
flask>=3.0.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0
python-dotenv==1.0.0
accelerate>=0.20.1